requirements.txt
pytz
pause
kartverket_tide_api
wifi
pybleno
rpi-ws281x